VMClone: Dont get to Failed phase in case of missing source
In case of missing source we dont wont to get to a terminal
Failed phase since the source can be created after the VMClone
was created. Update the conditions with matching conditions message.
Also instead of requeing because of the error we will exit the reconcile
and will be requeued once source is created.
This also managed to remove a redundant jump to updateStatus which is
against k8s design pattern.

Signed-off-by: Shelly Kagan <skagan@redhat.com>