VMSnapshot: mark volumesnapshotstatus disabled if pvc is unbound
K8S Volumesnapshot doesnt support taking a snapshot
when PVC is not bound, we should align the volumesnapshotstatus
disabled accordingly, this can help prevent users from taking a
VMSnapshot and not realizing that this unbound volume will not be
a part of the snapshot.
Changed the check of the storageclass accordingly, no interest in
checking the storageclassname on either the DV nor the DVTemplate
if the PVC doesnt exist/bound then this DV will not be snapshotable
either.
Adjusted UT accordingly and added a scenario of having a PVC unbound
hence the status should be `disabled`.

Signed-off-by: Shelly Kagan <skagan@redhat.com>