VMRestore: move update of dvname in restore
The update of the dvname on the vmrestore restores
list was done in the 'createDataVolume' function.
In case the creation of the DV succeeded but the
update of the vmrestore failed we would have never
gotten again to update the dvname.
Instead moved the update to be independent of the dv
creation and make sure the values are updated before continuing
to next step of the restore.
This should make the recently QUARANTINE tests
[https://github.com/kubevirt/kubevirt/pull/12824]
to pass, so removing the label and decorator.

Signed-off-by: Shelly Kagan <skagan@redhat.com>