Introduce VirtualMachineBackup and VirtualMachineBackupTracker CRDs
This CRDs will be used to initiate a full and incremental backups for VMs
- VirtualMachineBackupTracker is a namespace-scoped Custom Resource
designed to track the latest backup checkpoint per-VM, per-backup-solution basis.
It is intended to be created by a backup provider or user for a specific VM
to monitor and manage incremental backups associated with a given backup solution.
- VirtualMachineBackup is a namespace-scoped Custom Resource that initiates
the backup process.

Signed-off-by: Shelly Kagan <skagan@redhat.com>